Fix incorrect use of scoped enumerations in format strings
Scoped enums do not get automatically promoted to their underlying type,
so these uses have undefined behavior and Clang recently started warning
about it.

Bug: 1456289
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4632620
Commit-Queue: Nico Weber <thakis@chromium.org>
Owners-Override: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1160947}

+++ chrome/browser/ash/bruschetta/bruschetta_terminal_provider.cc
@@ -69,7 +69,8 @@ void BruschettaTerminalProvider::EnsureRunning(
               success ? ""
                       : base::StringPrintf(
                             "Error starting bruschetta for terminal: %d (%s)",
-                            result, BruschettaResultString(result)));
+                            static_cast<int>(result),
+                            BruschettaResultString(result)));
         },
         std::move(callback)));
   } else {
